+++ frontend/webapp/render/views/actions_view.py
@@ -1,5 +1,4 @@
-import json
-
+import ujson as json
 from flask import Blueprint
 from flask import current_app as app
 from flask import jsonify